+++ core/Common/src/main/java/de/markusbordihn/easynpc/client/renderer/entity/layers/EasyNPCItemAttachmentLayer.java
@@ -99,7 +99,7 @@ public class EasyNPCItemAttachmentLayer<
       return;
     }
 
-    EasyNPC<?> easyNPC = LivingEntityManager.getEasyNPCEntityByUUID(entityUUID);
+    EasyNPC<?> easyNPC = LivingEntityManager.getClientEasyNPCEntityByUUID(entityUUID);
     if (easyNPC == null) {
       return;
     }