+++ core/Forge/src/main/java/de/markusbordihn/easynpc/block/ModBlocks.java
@@ -128,9 +128,6 @@ public class ModBlocks {
                       .sound(SoundType.METAL)
                       .noOcclusion(),
                   SpawnerType.WORLD_SPAWNER));
-
-  private ModBlocks() {}
-
   public static final RegistryObject<BlockEntityType<EasyNPCSpawnerBlockEntityWrapper>>
       EASY_NPC_SPAWNER_ENTITY =
           BLOCK_ENTITY_TYPES.register(
@@ -144,4 +141,6 @@ public class ModBlocks {
                           EASY_NPC_SPAWNER_GROUP.get(),
                           EASY_NPC_SPAWNER_SINGLE.get(),
                           EASY_NPC_SPAWNER_WORLD.get())));
+
+  private ModBlocks() {}
 }